fix: ensure proper type casting and improve platform info handling
- Updated the tick method in `engine.cpp` to explicitly cast `m_delta_time` to float for type safety.
- Renamed variable `hKey` to `h_key` in `platform_info.cpp` for consistency and clarity.
- Enhanced command execution in `build_config.py` and `build.py` by adding optional encoding parameter to prevent character encoding issues on Windows.

+++ engine/engine.cpp
@@ -184,7 +184,7 @@ namespace RealmEngine
     {
         m_input->tick();
         m_window->pollEvents();
-        m_scene->getCurrentOrNewScene()->tick(m_delta_time);
+        m_scene->getCurrentOrNewScene()->tick(static_cast<float>(m_delta_time));
     }
 
     void Engine::renderTick()
